requirements.txt
gigachain
langchain-community==0.0.31
langchain_core
streamlit==1.32.2
sentence-transformers==2.6.1
faiss-cpu==1.8.0
python-dotenv
